+++ cmd/git-go/helpers.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"github.com/Nivl/git-go"
+	"github.com/Nivl/git-go/env"
 	"github.com/Nivl/git-go/internal/pathutil"
 )
 
@@ -15,12 +16,12 @@ func loadRepository(cfg *config) (*git.Repository, error) {
 
 	// run the command
 	return git.OpenRepositoryWithOptions(repoPath, git.OpenOptions{
-		EnvOptions: newEnvOptionsFromCfg(cfg),
+		GitOptions: newEnvOptionsFromCfg(cfg),
 	})
 }
 
-func newEnvOptionsFromCfg(cfg *config) *git.EnvOptions {
-	return &git.EnvOptions{
+func newEnvOptionsFromCfg(cfg *config) *env.GitOptions {
+	return &env.GitOptions{
 		GitDirPath:       cfg.GitDir,
 		GitObjectDirPath: cfg.GitObjectDir,
 		SkipSystemConfig: cfg.SkipSystemConfig,
