+++ cmd/git-go/init.go
@@ -20,7 +20,7 @@ func newInitCmd(cfg *config) *cobra.Command {
 
 func initCmd(cfg *config) error {
 	r, err := git.InitRepositoryWithOptions(cfg.C.String(), git.InitOptions{
-		EnvOptions: newEnvOptionsFromCfg(cfg),
+		GitOptions: newEnvOptionsFromCfg(cfg),
 	})
 	if err != nil {
 		return err
